+++ experimental/packages/opentelemetry-sdk-node/test/sdk.test.ts
@@ -131,10 +131,10 @@ describe('Node SDK', () => {
     });
 
     it('should not register more than the minimal SDK components', async () => {
-      // need to set these to none, since the deafult value is 'otlp'
+      // need to set these to none, since the default value is 'otlp'
       process.env.OTEL_TRACES_EXPORTER = 'none';
       process.env.OTEL_LOGS_EXPORTER = 'none';
-      process.env.OTEL_METRIC_EXPORTER = 'none';
+      process.env.OTEL_METRICS_EXPORTER = 'none';
       const sdk = new NodeSDK({
         autoDetectResources: false,
       });
@@ -201,8 +201,6 @@ describe('Node SDK', () => {
 
       sdk.start();
 
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
-
       assertDefaultContextManagerRegistered();
       assertDefaultPropagatorRegistered();
 
@@ -219,8 +217,6 @@ describe('Node SDK', () => {
       });
 
       sdk.start();
-
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
 
       assertDefaultContextManagerRegistered();
       assertDefaultPropagatorRegistered();
@@ -244,8 +240,6 @@ describe('Node SDK', () => {
       });
 
       sdk.start();
-
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
 
       assertDefaultContextManagerRegistered();
       assertDefaultPropagatorRegistered();
@@ -1246,13 +1240,6 @@ describe('Node SDK', () => {
       delete process.env.OTEL_EXPORTER_METRICS_PROTOCOL;
     });
 
-    it('should not register the provider if OTEL_METRICS_EXPORTER is not set', async () => {
-      const sdk = new NodeSDK();
-      sdk.start();
-      assert.ok(!(metrics.getMeterProvider() instanceof MeterProvider));
-      await sdk.shutdown();
-    });
-
     it('should not register the provider if OTEL_METRICS_EXPORTER contains none', async () => {
       process.env.OTEL_METRICS_EXPORTER = 'console,none';
       const sdk = new NodeSDK();
@@ -1466,7 +1453,7 @@ describe('Node SDK', () => {
     const sharedState = (meterProvider as any)['_sharedState'];
     assert.ok(
       sharedState.metricCollectors[0]._metricReader._exporter instanceof
-      OTLPGrpcMetricExporter
+        OTLPProtoMetricExporter
     );
     await sdk.shutdown();
   });